fixing shit et ajout des boutton retour

+++ app/src/main/java/com/example/app_miel/data/Data_miel.java
@@ -3,11 +3,11 @@ package com.example.app_miel.data;
 
 public class Data_miel {
 
-    private int     id_miel;
+    private Integer     id_miel;
     private String  nom_miel;
-    private int     prix_miel;
+    private Float     prix_miel;
 
-    public int getId_miel() {
+    public Integer getId_miel() {
         return id_miel;
     }
 
@@ -15,7 +15,7 @@ public class Data_miel {
         return nom_miel;
     }
 
-    public int getPrix_miel() {
+    public Float getPrix_miel() {
         return prix_miel;
     }
 }
